[nrf fromtree] tests: Bluetooth: df: Move common code to separate ...
source file

Move common code that may be used by different test units
from main.c to common.c. Main.c should be clean and responsible
for unit tests executioon only. Common functionalities should be
stored in separate file and provided to test source files by
appropriate header file.
